opc_ua_system/requirements.txt
opcua-asyncio>=1.1.0
asyncua>=1.0.0
torch>=2.1.0
numpy>=1.24.0
scikit-learn>=1.3.0
neo4j>=5.14.0
rdflib>=7.0.0
owlrl>=6.0.0
pandas>=2.1.0
tqdm>=4.66.0
matplotlib>=3.8.0
pyyaml>=6.0
lxml>=5.0.0
jupyter>=1.0.0
pytorch-crf>=0.7.0
